New law change type possible on ion storm
A new law change type that takes a certain law and appends stuff to it, like not being able to tell the law etc. Comes with an Addition risk mode. This proc needs to be implemented in the /datum/round_event/ion_storm/start() found in code/modules/events/ion_storm.dm, which isn't modular yet.